sif_vaskelister/pyproject.toml
[tool.poetry]
name = "sif_vaskelister"
version = "0.1.0"
description = "A web cleaning management system"
authors = ["Aksel Lenes <aksel.lenes@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.3"
django-oauth-toolkit = "^1.2.0"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = { version = "^4.3.21", extras = ["pyproject"] }
pyyaml = "^5.3"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
known_django = "django"
sections = [ "FUTURE",
             "STDLIB",
             "DJANGO",
             "THIRDPARTY",
             "FIRSTPARTY",
             "LOCALFOLDER" ]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
